crypto/requirements.txt
web3==6.0.0
python-dotenv==0.20.0
requests==2.28.1
telegram==13.7
google-auth==2.20.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
schedule==1.1.0